requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
fastapi==0.115.3
h11==0.14.0
idna==3.10
Mako==1.3.6
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
python-env==1.0.0
python-multipart==0.0.17
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlmodel==0.0.22
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
